Move keys logic into machine
It's not very helpful as a separate class.

Also, fix the behavior to match Chip8: the await condition isn't satisfied
until the first key that goes down also comes back up.

See: https://www.laurencescotford.net/2020/07/19/chip-8-on-the-cosmac-vip-keyboard-input/

There should also be a beep while the key is down, but it's not as important.